requirements.txt
alabaster==0.7.12
asgiref==3.3.4
astroid==2.5.6
Babel==2.9.1
beautifulsoup4==4.10.0
better-apidoc==0.3.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2
django-bootstrap4==3.0.1
django-filter==21.1
django-formset-js==0.5.0
django-jquery-js==3.1.1
django-smart-selects==1.5.9
django-widget-tweaks==1.4.8
docutils==0.16
html5lib==1.1
idna==2.10
imagesize==1.2.0
isort==5.8.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.8.2
pyparsing==2.4.7
Pyphen==0.10.0
python-dotenv==0.17.1
pytz==2021.1
pyxdg==0.27
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-fulltoc==1.2.0
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.4
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xdg==5.0.2
